Reimplemented the whole Dialog system.
The dialog is now a list of all dialog screens.
At the bottom of the dialog you get an option to continue to the next screen.
If there is no next screen the option given is to exit the dialog and go back to the previous state (i.e. zone).

TODO:
- Give an option for the dialog to be a quest, and for the player to accept or deny it.
- No repetitive dialogs.